Remake the old main in a proper architecture
